solid_cache stores BlockID instead of bool — eliminates redundant get_block_unsafe calls
solid_cache now stores the actual uint16_t BlockID for every position
(both interior and chunk-boundary entries). The center-block get_block_unsafe
call is eliminated in both greedy_h and greedy_v — read straight from cache.
In greedy_v, ALL four side-neighbor lookups are replaced by cache reads,
removing the boundary-check switch and per-neighbor get_block_unsafe calls.
In greedy_h, the fast-path neighbor-above check also reads from cache.
The non-greedy fallback path in mesh_builder.cpp is similarly simplified.

Estimated savings: one get_block_unsafe (palette traversal) eliminated per
non-air voxel in greedy_h, and up to five (center + 4 neighbors) in greedy_v.

+++ src/mesh/mesh_builder.hpp
@@ -259,8 +259,10 @@ private:
     std::vector<uint32_t> water_indices;
     // Layout is [y][z][x] (not [x][z][y]) so that scans with x as the inner
     // loop variable (passive_greedy_mesh_horizontal) read sequential memory
-    // instead of striding by SC_W * SC_D each step
-    std::array<std::array<std::array<uint8_t, SC_W>, SC_D>, CHUNK_HEIGHT> solid_cache{};
+    // instead of striding by SC_W * SC_D each step.
+    // Stores BlockID of the block at chunk position (x-1, y, z-1) for interior
+    // entries, or the neighbor-chunk block for boundary entries.
+    std::array<std::array<std::array<uint16_t, SC_W>, SC_D>, CHUNK_HEIGHT> solid_cache{};
 
     // Neighbor chunk accessor for cross-chunk block/light access.
     ChunkNeighborAccessor accessor;
